Extract custom config logic to a file
